+++ src/com/gn/module/partner/PartnerController.java
@@ -1,6 +1,7 @@
 package com.gn.module.partner;
 
 import com.gn.database.DbUtil;
+import com.gn.model.Member;
 import com.gn.model.Partner;
 import com.gn.model.TableData;
 import com.gn.module.dialog.DialogPartner;
@@ -12,9 +13,7 @@ import javafx.fxml.FXMLLoader;
 import javafx.fxml.Initializable;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
-import javafx.scene.control.TableColumn;
-import javafx.scene.control.TableRow;
-import javafx.scene.control.TableView;
+import javafx.scene.control.*;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.stage.Modality;
 import javafx.stage.Stage;
@@ -95,8 +94,13 @@ public class PartnerController implements Initializable {
 
     @FXML
     public void deletePartner() {
-        Partner selectedRow = tableView.getSelectionModel().getSelectedItem();
-        dbUtil.deletePartner(selectedRow.getPartnerId());
-        updateView();
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION, "Bạn có chắc chắn muốn xóa?", ButtonType.YES, ButtonType.NO);
+        alert.showAndWait();
+
+        if (alert.getResult() == ButtonType.YES) {
+            Partner selectedRow = tableView.getSelectionModel().getSelectedItem();
+            dbUtil.deletePartner(selectedRow.getPartnerId());
+            updateView();
+        }
     }
 }
